Big changes model improve UX and avoid complexity
