--- project/src/Entity/Room.php
@@ -1,106 +1,0 @@
-<?php
-
-namespace App\Entity;
-
-use App\Repository\RoomRepository;
-use Doctrine\Common\Collections\ArrayCollection;
-use Doctrine\Common\Collections\Collection;
-use Doctrine\DBAL\Types\Types;
-use Doctrine\ORM\Mapping as ORM;
-
-#[ORM\Entity(repositoryClass: RoomRepository::class)]
-class Room
-{
-    #[ORM\Id]
-    #[ORM\GeneratedValue]
-    #[ORM\Column]
-    private ?int $id = null;
-
-    #[ORM\Column(length: 30)]
-    private ?string $name = null;
-
-    #[ORM\Column(type: Types::TEXT, nullable: true)]
-    private ?string $description = null;
-
-    #[ORM\Column(nullable: true)]
-    private ?bool $reserved = null;
-
-    #[ORM\OneToMany(mappedBy: 'room', targetEntity: Borrow::class)]
-    private Collection $borrows;
-
-    public function __construct()
-    {
-        $this->borrows = new ArrayCollection();
-    }
-
-    public function getId(): ?int
-    {
-        return $this->id;
-    }
-
-    public function getName(): ?string
-    {
-        return $this->name;
-    }
-
-    public function setName(string $name): self
-    {
-        $this->name = $name;
-
-        return $this;
-    }
-
-    public function getDescription(): ?string
-    {
-        return $this->description;
-    }
-
-    public function setDescription(?string $description): self
-    {
-        $this->description = $description;
-
-        return $this;
-    }
-
-    public function isReserved(): ?bool
-    {
-        return $this->reserved;
-    }
-
-    public function setReserved(?bool $reserved): self
-    {
-        $this->reserved = $reserved;
-
-        return $this;
-    }
-
-    /**
-     * @return Collection<int, Borrow>
-     */
-    public function getBorrows(): Collection
-    {
-        return $this->borrows;
-    }
-
-    public function addBorrow(Borrow $borrow): self
-    {
-        if (!$this->borrows->contains($borrow)) {
-            $this->borrows->add($borrow);
-            $borrow->setRoom($this);
-        }
-
-        return $this;
-    }
-
-    public function removeBorrow(Borrow $borrow): self
-    {
-        if ($this->borrows->removeElement($borrow)) {
-            // set the owning side to null (unless already changed)
-            if ($borrow->getRoom() === $this) {
-                $borrow->setRoom(null);
-            }
-        }
-
-        return $this;
-    }
-}
